requirements.txt
django==3.1
djangorestframework==3.11.1
markdown==3.0
django-extensions
